AMB-2181- Non-sync file-permission-updates

+++ recordforwarder/src/forwarding_lambda.py
@@ -39,8 +39,9 @@ def identify_supplier(file_key: str):
     return ODS_PATTERNS.get(ods_code, None)
 
 
-def forward_request_to_api(message_header, bucket_name, file_key, action_flag, fhir_json, ack_bucket_name, imms_id,
-                           version):
+def forward_request_to_api(
+    message_header, bucket_name, file_key, action_flag, fhir_json, ack_bucket_name, imms_id, version
+):
     response = s3_client.head_object(Bucket=bucket_name, Key=file_key)
     created_at = response["LastModified"]
     created_at_formatted = created_at.strftime("%Y%m%dT%H%M%S00")
@@ -64,42 +65,45 @@ def forward_request_to_api(message_header, bucket_name, file_key, action_flag, f
             # File doesn't exist, write the header to the new file
             accumulated_csv_content.write("|".join(headers) + "\n")
             print(f"accumulated_csv_content:{accumulated_csv_content}")
-    if imms_id == "None" and version == "None":
-        data_row = Constant.data_rows("None", created_at_formatted, message_header)
-    supplier_system = identify_supplier(file_key)
-    if action_flag == "new":
-        response, status_code = immunization_api_instance.create_immunization(fhir_json, supplier_system)
-        print(f"response:{response},status_code:{status_code}")
-        if status_code == 201:
-            data_row = Constant.data_rows(True, created_at_formatted, message_header)
-        elif status_code == 422:
-            data_row = Constant.data_rows("duplicate", created_at_formatted, message_header)
-        else:
-            data_row = Constant.data_rows(False, created_at_formatted, message_header)
-    elif action_flag == "update" and imms_id not in (None, "None") and version not in (None, "None"):
-        fhir_json["id"] = imms_id
-        print(f"updated_fhir_json:{fhir_json}")
-        response, status_code = immunization_api_instance.update_immunization(
-            imms_id, version, fhir_json, supplier_system
-        )
-        print(f"response:{response},status_code:{status_code}")
-        if status_code == 200:
-            data_row = Constant.data_rows(True, created_at_formatted, message_header)
-        else:
-            data_row = Constant.data_rows(False, created_at_formatted, message_header)
-    elif action_flag == "delete" and imms_id not in (None, "None"):
-        response, status_code = immunization_api_instance.delete_immunization(imms_id, fhir_json, supplier_system)
-        print(f"response:{response},status_code:{status_code}")
-        if status_code == 204:
-            data_row = Constant.data_rows(True, created_at_formatted, message_header)
-        else:
-            data_row = Constant.data_rows(False, created_at_formatted, message_header)
+    if fhir_json == "No_Permissions":
+        data_row = Constant.data_rows("no permissions", created_at_formatted, message_header)
+    else:
+        if imms_id == "None" and version == "None":
+            data_row = Constant.data_rows("None", created_at_formatted, message_header)
+        supplier_system = identify_supplier(file_key)
+        if action_flag == "new":
+            response, status_code = immunization_api_instance.create_immunization(fhir_json, supplier_system)
+            print(f"response:{response},status_code:{status_code}")
+            if status_code == 201:
+                data_row = Constant.data_rows(True, created_at_formatted, message_header)
+            elif status_code == 422:
+                data_row = Constant.data_rows("duplicate", created_at_formatted, message_header)
+            else:
+                data_row = Constant.data_rows(False, created_at_formatted, message_header)
+        elif action_flag == "update" and imms_id not in (None, "None") and version not in (None, "None"):
+            fhir_json["id"] = imms_id
+            print(f"updated_fhir_json:{fhir_json}")
+            response, status_code = immunization_api_instance.update_immunization(
+                imms_id, version, fhir_json, supplier_system
+            )
+            print(f"response:{response},status_code:{status_code}")
+            if status_code == 200:
+                data_row = Constant.data_rows(True, created_at_formatted, message_header)
+            else:
+                data_row = Constant.data_rows(False, created_at_formatted, message_header)
+        elif action_flag == "delete" and imms_id not in (None, "None"):
+            response, status_code = immunization_api_instance.delete_immunization(imms_id, fhir_json, supplier_system)
+            print(f"response:{response},status_code:{status_code}")
+            if status_code == 204:
+                data_row = Constant.data_rows(True, created_at_formatted, message_header)
+            else:
+                data_row = Constant.data_rows(False, created_at_formatted, message_header)
 
     data_row_str = [str(item) for item in data_row]
     cleaned_row = "|".join(data_row_str).replace(" |", "|").replace("| ", "|").strip()
 
     accumulated_csv_content.write(cleaned_row + "\n")
-
+    print(f"CSV content before upload:\n{accumulated_csv_content.getvalue()}")
     # Upload the updated CSV content to S3
     csv_file_like_object = io.BytesIO(accumulated_csv_content.getvalue().encode("utf-8"))
     s3_client.upload_fileobj(csv_file_like_object, ack_bucket_name, ack_filename)
@@ -124,8 +128,10 @@ def forward_lambda_handler(event, context):
             if action_flag in ("update", "delete", "None"):
                 imms_id = message_body.get("imms_id")
                 version = message_body.get("version")
-            forward_request_to_api(message_header, bucket_name, file_key, action_flag, fhir_json, ack_bucket_name,
-                                   imms_id, version)
+
+            forward_request_to_api(
+                message_header, bucket_name, file_key, action_flag, fhir_json, ack_bucket_name, imms_id, version
+            )
 
         except Exception as e:
             logger.error(f"Error processing message: {e}")
